Sort by version_numified, status (prefer CPAN) then date.
The original code was working around the rounding score issue at
https://github.com/CPAN-API/cpan-api/issues/284

while the bug isn't fixed as of today yet, the change of logic in sort
order whether it's dev or not doesn't sound right.

For example, Plack::Middleware has always been set to VERSION undef,
and MetaCPAN search right now returns them in a random
order. Preferring the oldest distribution for that definitely isn't
the right thing to do.

For a module with VERSION set to non-zero value, MetaCPAN only seem to
index them when the value increases, just like PAUSE, so if the
distribution is returned in the response it's OK to handle the latest
one.